refactor(OneDrive Node): Tidy trigger scope guard and harness fixture path
- Hoist the trigger LastLink scope check to an optional-chained
  hasValidScope const (prefer-optional-chain).
- Point the harness tests at the relocated test/credentials.ts fixture
  (../credentials) now that it lives under the test tree.
